Question 2

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -5,9 +5,9 @@
  *
  * Return length of string
  */
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
-	size_t length = 0;
+	int length = 0;
 	while (*str++)
 		length++;
